pyproject.toml
[project]
name = "thuc-tap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.7",
    "ipython>=9.12.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.1.6",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.22",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "google-generativeai>=0.8.3",
    "geopy>=2.4.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.31.0",
    "requests>=2.32.0",
    "watchfiles>=1.0.0",
    "gspread>=6.1.4",
    "oauth2client>=4.1.3",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
]
